Move allocCopy functions to NGAP utilities
The file ngap_gNB_management_procedures.c is containing
NGAP utility functions. The allocCopy function was locally defined
in different files to perform memory copy to/from the NGAP
PDU to/from OCTET STRING.

Since the same functions are needed in the N2 handover procedures,
this commit aims at improving the maintainability of the code base
by moving the functions to a common utility file. For a better
readability of the code, the functions are also renamed,
depending on the direction of the copy, i.e. to/from Octet String.

Finally, the new functions are adopted in stack.