Rework handover ack callback
Pass the xid only, instead of buffer and size, and call RRCReconfiguration
message generation inside the F1 callback. This will semplify the code in the
rrc_CU_process_ue_context_setup_response function once the N2 handover case
is introduced. In rrc_CU_process_ue_context_setup_response, with presence
of the HO context, the HO ACK callback is called only once and depending
on the type of HO it will do either RRCReconfiguration for F1 and
HandoverCommand for N2.