UE changes: Handling of ulsyncvalidityduration in sib19
Spec 38.331 section 5.2.2.4.21 indicates that T430 started with duration in ulsyncvalidityduration in SIB19
from epochtime. If epochtime is not sent, it is started when sib19 is received by UE.
Implementation related to epochtime is still under discussion.
Spec 38.331 section 5.2.2.6 indicates that after T430 expiry (if SIB19 not received in time) , UL SYNC needs
to be re-obtained by performing Randome access to get the UL Timing advance.