Refactor rrc_gNB_encode_RRCReconfiguration
Add a function to prepare RRCReconfigurationParams for RRC encoding.

Depending on the RRCReconfiguration message to be encoded, content of IEs
can differ, e.g. PDCP reestablishment can differ.

The function is preparing the params for the IEs that are
either mandatory or shared among different RRC messages. The input
provides the flag/bitmap for PDCP re-establishment of DRBs/SRBs.

Specific fields that depend on the message to be encoded are then
added afterwards in the caller.

Finally, the encoded message coming from rrc_gNB_encode_RRCReconfiguration
is passed to the handover callback and based on the type of handover it is used
either to trigger a new RRCReconfiguration (F1) or is passed in a
container of an NGAP message (N2, Handover Preparation, Handover Command)